requirements.txt
certifi==2017.11.5
chardet==3.0.4
ChatterBot==0.8.1
chatterbot-corpus==1.1.1
chatterbot-weather==0.1.1
cookies==2.2.1
future==0.16.0
idna==2.6
jsondatabase==0.1.7
mathparse==0.1.1
nltk==3.2.5
oauthlib==2.0.6
pymongo==3.6.0
python-dateutil==2.6.1
python-forecastio==1.3.4
python-twitter==3.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
responses==0.8.1
six==1.11.0
slackclient==1.1.0
SQLAlchemy==1.1.15
urllib3==1.22
websocket-client==0.44.0
